Add verification preference support
Post-upload program verification can be controlled via File >
Preferences > Verify code after upload in Arduino IDE 1.6.9 and newer.
Default values of the upload.verify and program.verify properties must
be set in order for an AVRDUDE command to be generated in Arduino IDE
1.6.8 and older that is compatible with AVRDUDE 6.3.0.

MicroCore's tools.avrdude.program items are only used when one of the
(slow) programmer options are selected. This means that Upload Using
Programmer will still not be compatible with Arduino IDE 1.6.8 or older
using Arduino AVR Boards 1.6.12 or 1.6.14 when one of the standard Tools
> Programmer menu selections are used since the platform.txt associated
with the selected programmer is used for Upload Using Programmer.